+++ apigen/www/index.php
@@ -1,437 +1,88 @@
-<?php
-    require_once("../server/bootstrap.php");
-?>
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" lang="en" >
-<head>
-<script src="//code.jquery.com/jquery-1.12.4.min.js" integrity="sha256-ZosEbRLbNQzLpnKIkEdrPv7lOy9C27hHQ+Xp8a4MxAQ=" crossorigin="anonymous"></script>
-<title>Web Framework</title>
-<script>
+<?php require_once("top.inc.php"); ?>
 
-    var new_category_form_visible = false;
-    var edit_category_form_visible = false;
-
-    function showEditCategoryForm(nombre,descripcion,id)
-    {
-        if(!edit_category_form_visible)
-        {
-            edit_category_form_visible = true;
-
-            var html = '';
-            html += '<div id="ec">'
-            html += '<label> Nombre </label>';
-            html += '<input type="text" name="nombre_clasificacion" id="nombre_clasificacion_edit" value="'+nombre+'">';
-            html += '<label> Descripcion </label>';
-            html += '<textarea name="descripcion_clasificacion" id="descripcion_clasificacion_edit">'+descripcion+'</textarea>';
-            html += '<input type="hidden" name="id_proyecto" value="<?php if(isset($_GET["project"]) && is_numeric($_GET["project"])) echo $_GET["project"]; ?>">';
-            html += '<input type="hidden" name="id_clasificacion" value='+id+'>';
-            html += '<input type="button" onClick="validarCamposCategoriaEditar()" value="Enviar">';
-            html += '<a onClick="hideEditCategoryForm()">Hide</a>';
-            html += '</div>';
-
-            $("#editar_categoria").append(html);
-
-        }
-    }
-
-    function validarCamposCategoriaEditar()
-    {
-            if($.trim($("#nombre_clasificacion_edit").val())=="")
-            {
-                alert("Falta el nombre de la clasificacion");
-                return;
-            }
-            if($.trim($("#nombre_clasificacion_edit").val()).search(/['"]+/)>=0)
-            {
-                alert("El nombre de la clasificacion tiene comillas simples o dobles, use en su lugar `");
-                return;
-            }
-            if($.trim($("#descripcion_clasificacion_edit").val()).search(/['"]+/)>=0)
-            {
-                alert("La descripcion de la clasificacion tiene comillas simples o dobles, use en su lugar `");
-                return;
-            }
-            $("#editar_categoria").submit();
-    }
-
-    function hideEditCategoryForm()
-    {
-        if(edit_category_form_visible)
-            {
-                edit_category_form_visible = false;
-
-                $("#ec").remove();
-            }
-    }
-
-    function showNewCategoryForm()
-    {
-        if(!new_category_form_visible)
-        {
-
-            new_category_form_visible = true;
-
-            var html = '';
-            html += '<div id="nc">'
-            html += '<label> Nombre </label>';
-            html += '<input type="text" name="nombre_clasificacion" id="nombre_clasificacion">';
-            html += '<label> Descripcion </label>';
-            html += '<textarea name="descripcion_clasificacion" id="descripcion_clasificacion"></textarea>';
-            html += '<input type="hidden" name="id_proyecto" value="<?php if(isset($_GET["project"]) && is_numeric($_GET["project"])) echo $_GET["project"]; ?>">';
-            html += '<input type="button" onClick="validarCamposCategoria()" value="Enviar">';
-            html += '<a onClick="hideNewCategoryForm()">Hide</a>';
-            html += '</div>';
-
-            $("#nueva_categoria").append(html);
-
-        }
-    }
-
-    function validarCamposCategoria()
-    {
-            if($.trim($("#nombre_clasificacion").val())=="")
-            {
-                alert("Falta el nombre de la clasificacion");
-                return;
-            }
-            if($.trim($("#nombre_clasificacion").val()).search(/['"]+/)>=0)
-            {
-                alert("El nombre de la clasificacion tiene comillas simples o dobles, use en su lugar `");
-                return;
-            }
-            if($.trim($("#descripcion_clasificacion").val()).search(/['"]+/)>=0)
-            {
-                alert("La descripcion de la clasificacion tiene comillas simples o dobles, use en su lugar `");
-                return;
-            }
-            $("#nueva_categoria").submit();
-    }
-
-    function hideNewCategoryForm()
-    {
-        if(new_category_form_visible)
-            {
-                new_category_form_visible = false;
-
-                $("#nc").remove();
-            }
-    }
-
-      function Borrar(id)
-      {
-         var selection = confirm("Esta seguro de querer borrar el método con todos sus argumentos y repsuestas?");
-
-         if(selection)
-            window.location="delete_method.php?m="+id<?php if(isset($_GET["cat"])) echo '+"&cat='.$_GET["cat"].'"'; if(isset($_GET["project"])) echo '+"&project='.$_GET["project"].'"'?>;
-      }
-
-      function ProjectChange(val)
-      {
-          window.location = "index.php?project="+val;
-      }
-
-      function Borrar_categoria()
-      {
-         var selection = confirm("Esta seguro de querer borrar la categoria con todos su metodos?");
-
-         if(selection)
-            window.location="delete_cat.php?"<?php if(isset($_GET["cat"])) echo '+"&cat='.$_GET["cat"].'"'; if(isset($_GET["project"])) echo '+"&project='.$_GET["project"].'"'?>;
-      }
-
-</script>
-
-<link type="text/css" rel="stylesheet" href="media/f.css"/>
-
-</head>
-<body class="safari4 mac Locale_en_US">
-<input type="hidden" autocomplete="off" id="post_form_id" name="post_form_id" value="d8f38124ed9e31ef3947198c6d26bff1"/>
-<div id="FB_HiddenContainer" style="position:absolute; top:-10000px; width:0px; height:0px;">
-</div>
 <div class="devsitePage">
-    <div class="menu">
-        <div class="content">
+
+<?php require_once("menu.inc.php"); ?>
+
+<div class="body nav">
+    <div class="content">
+
+        <div id="bodyMenu" class="bodyMenu">
+        <?php require_once("toplevnav.inc.php"); ?>
+        </div>
+
+            <div id="bodyText" class="bodyText">
+
+
+            <?php require_once("titleheader.inc.php"); ?>
+
 
             <?php
-            $proyecto = null;
-            if(isset($_GET["project"]) && is_numeric($_GET["project"]))
+            //if(isset($_GET["project"]) && is_numeric($_GET["project"]))
             {
-                $proyecto = mysql_fetch_assoc(mysql_query(" Select * from mantis_project_table where id =".$_GET["project"]));
-            }
+                // If a looking at a category or method, expand the methods
+                if(!isset($_GET["m"]) && !isset($_GET["cat"])){
+                    $query = mysql_query("select * from clasificacion where id_proyecto = ". $_GET["project"]." order by nombre asc;");
 
-            if (isset($_GET["m"])) {
-                echo '<a class="l" href="em.php?m='. $_GET["m"] .'&cat='.$_GET["cat"].'&project='.$_GET["project"].'">Editar</a>';
-                echo '<a class="l" onClick="Borrar('. $_GET["m"] .')">Borrar</a>';
-
-            }
-
-
-            if(isset($_GET["project"])&&  is_numeric($_GET["project"]))
-            {
-                echo '<a class="l" href="nm.php?project='.$_GET["project"];
-                    if(isset($_GET["cat"])) 
-                        echo "&cat=".$_GET["cat"]; 
-                echo '">Nuevo metodo</a> ';
-            }
-
-            if(isset($_GET["project"])&&  is_numeric($_GET["project"]))
-            {
-                ?>
-                    <a class="l" href="build.php<?php echo '?project='.$_GET["project"] ?>">Generar</a>
-                <?php
-            }
-            ?>
-
-            <a class="l">Proyecto: 
-
-            <select name="project" id="project" onChange = "ProjectChange(this.value)" >
-                <option value = "null"> ------------ </option>
-                <?php
-
-                $query = "select id as id_proyecto,name as nombre from mantis_project_table";
-                $res = mysql_query($query);
-                while($row = mysql_fetch_assoc($res))
-                {
-                    if(isset($_GET["project"]) && $_GET["project"] == $row["id_proyecto"])
+                    while( ($row = mysql_fetch_assoc( $query )) != null )
                     {
-                        echo "<option value = ".$row["id_proyecto"]." selected>".$row["nombre"]."</option>";
-                    }
-                    else
-                    {
-                        echo "<option value = ".$row["id_proyecto"].">".$row["nombre"]."</option>";
-                    }
-                }
-
-                ?>
-            </select>
-
-            </a>
-            <div class="clear">
-            </div>
-        </div>
-    </div>
-
-    <div class="body nav">
-        <div class="content">
-            <div id="bodyMenu" class="bodyMenu">
-                <div class="toplevelnav">
-
-                                    <?php if(isset($_GET["project"]) && is_numeric($_GET["project"]))
-                                    {
-                                    ?>
-                    <div id="form_nueva_categoria">
-                                            <a onClick="showNewCategoryForm()">Nueva categoria</a>
-                                            <form id="nueva_categoria" method="POST" action="negocios_clasificacion.php">
-
-                                            </form>
-                                        </div>
-                                    <?php } ?>
-
-                                    <ul>
-                        <?php
-
-                        if(isset($_GET["project"]) && is_numeric($_GET["project"]) )
-                        {
-                            $query = mysql_query("select * from clasificacion where id_proyecto=".$_GET["project"]." order by nombre ;");
-
-
-                            while( ($row = mysql_fetch_assoc( $query )) != null )
-                            {
-                                if(isset($_GET["cat"]) && ($_GET["cat"] == $row["id_clasificacion"]) ){
-                                    ?>
-                                        <li class="active withsubsections">
-                                        <a class="selected" href="index.php?cat=<?php echo $row["id_clasificacion"]; ?>&project=<?php echo $_GET["project"]?>">
-                                        <div class="navSectionTitle">
-                                            <?php echo $row["nombre"]; ?>
-                                        </div>
-                                        </a>
-                                        <div id="form_editar_categoria">
-                                            <a onClick="showEditCategoryForm(<?php echo "'".$row["nombre"]."','".$row["descripcion"]."',".$row["id_clasificacion"]?>)">Editar categoria</a>
-                                            <form id="editar_categoria" method="POST" action="negocios_clasificacion_editar.php"> </form>
-                                        </div>
-                                            <div id="borrar_categoria">
-                                                <a onClick="Borrar_categoria();">Borrar categoria</a>
-                                            </div>
-                                        <ul class="subsections">
-
-                                        <?php
-                                        $argsq = mysql_query("select * from metodo where id_clasificacion = ". $row["id_clasificacion"] ." order by nombre;");
-
-                                        while(($m = mysql_fetch_assoc($argsq)) != null)
-                                        {
-
-                                                $n = str_replace("api/", "", $m["nombre"] );
-                                                $n = substr(  $n , strpos( $n , "/" ) +1 );
-                                                echo '<li><a href="?&cat='.$row["id_clasificacion"].'&m='.$m["id_metodo"].'&project='.$_GET["project"].'">' . $n .  '</a></li>';
-                                        }
-                                        ?>
-                                        </ul>
-                                        </li>
-                                        <?php
-
-                                    }else{
-
-                                        ?>
-                                        <li>
-                                        <a href="index.php?cat=<?php echo $row["id_clasificacion"]; ?>&project=<?php echo $_GET["project"] ?>">
-                                            <div class="navSectionTitle">
-                                            <?php echo $row["nombre"]; ?>
-                                            </div>
-                                        </a>
-                                        </li>
-                                        <?php   
-                                    }
-
-                                }
-                                                            }
                         ?>
-
-
-                    </ul>
-                </div>
-
-                <ul id="navsubsectionpages">
-
-                </ul>
-            </div>
-            <div id="bodyText" class="bodyText">
-                <div class="header">
-                    <div class="content">
-
-                        <?php
-                            if(isset($_GET["mensaje"])){
-                                echo "<h3>" . $_GET["mensaje"] . "</h3>";
-                            }
-                            if(isset($_GET["m"])){
-                                $res = mysql_query("select * from metodo where id_metodo = " . $_GET["m"]) or die(mysql_error());
-                                $metodo = mysql_fetch_assoc($res);
-                                echo "<h1>" .$metodo["tipo"] . " " . $metodo["nombre"] . "</h1>";
-
-                            }else if(isset($_GET["cat"])){
-                                $res = mysql_query("select * from clasificacion where id_clasificacion = " . $_GET["cat"]) or die(mysql_error());
-                                $metodo = mysql_fetch_assoc($res);
-                                echo "<h1>" . $metodo["nombre"] . "</h1>";
-
-                            }
-                        ?>
-
-                        <div class="breadcrumbs">
-                                                    <?php
-                                                    if(!is_null($proyecto))
-                                                    {
-                            echo '<a href=".?project='.$_GET["project"].'">'.$proyecto["name"].'</a><br> ';
-                                                        echo $proyecto["description"];
-                                                    }
-
-                            if(isset($_GET["cat"])){
-                                $res = mysql_query("select * from clasificacion where id_clasificacion = " . $_GET["cat"]) or die(mysql_error());
-                                $metodo = mysql_fetch_assoc($res);
-
-                                echo'<br>&rsaquo; <a href=".?project='.$_GET["project"].'">'  . $metodo["nombre"] .  '</a>';
-                            }
-                            ?>
-
+                        <li class="active withsubsections">
+                        <a class="selected" href="index.php?cat=<?php echo $row["id_clasificacion"]; echo "&project=".$_GET["project"] ?> ">
+                        <div class="navSectionTitle">
+                            <?php echo $row["nombre"]; ?>
                         </div>
-                            <?php
-                            if(isset($_GET["cat"]) && !isset($_GET["m"])){
-                                $res = mysql_query("select * from clasificacion where id_clasificacion = " . $_GET["cat"]) or die(mysql_error());
-                                $metodo = mysql_fetch_assoc($res);
-
-                                echo "<p>" . $metodo["descripcion"] . "</p>";
-                            }
-                            ?>
-
-                    </div>
-                </div>
-
-<!-- ----------------------------------------------------------------------
-     ---------------------------------------------------------------------- -->
+                        </a>
+                        <ul class="subsections">
 
                         <?php
-                                                if(isset($_GET["project"]) && is_numeric($_GET["project"]))
-                                                {
-                            if(!isset($_GET["m"]) && !isset($_GET["cat"])){
-                                $query = mysql_query("select * from clasificacion where id_proyecto = ".$_GET["project"]." order by nombre asc;");
+                        $argsq = mysql_query("select * from metodo where id_clasificacion = ". $row["id_clasificacion"] ." order by nombre;");
 
-                                while( ($row = mysql_fetch_assoc( $query )) != null )
-                                {
-
-                                    ?>
-                                    <li class="active withsubsections">
-                                    <a class="selected" href="index.php?cat=
-                                                                        <?php echo $row["id_clasificacion"]; 
-                                                                            echo "&project=".$_GET["project"] 
-                                                                        ?>
-                                                                           ">
-                                    <div class="navSectionTitle">
-                                        <?php echo $row["nombre"]; ?>
-                                    </div>
-                                    </a>
-                                    <ul class="subsections">
-
-                                    <?php
-                                    $argsq = mysql_query("select * from metodo where id_clasificacion = ". $row["id_clasificacion"] ." order by nombre;");
-
-                                    while(($m = mysql_fetch_assoc($argsq)) != null)
-                                    {
-
-
-                                        echo '<li><a href="?&cat='.$row["id_clasificacion"].'&m='.$m["id_metodo"].'&project='.$_GET["project"].'">' . $m["nombre"] .  '</a></li>';
-                                    }
-                                    ?>
-                                    </ul>
-                                    </li>
-                                    <?php
-
-                                }
-                            }
+                        while(($m = mysql_fetch_assoc($argsq)) != null)
+                        {
+                            echo '<li><a href="?&cat='.$row["id_clasificacion"].'&m='.$m["id_metodo"].'&project='.$_GET["project"].'">' . $m["nombre"] .  '</a></li>';
                         }
                         ?>
+                        </ul>
+                        </li>
+                        <?php
+
+                    }
+                }
+            }
+            ?>
                 <p>
                     <?php
+                    // Method description
                     if(isset($_GET["m"])){
                         $res = mysql_query("select * from metodo where id_metodo = " . $_GET["m"]) or die(mysql_error());
                         $metodo = mysql_fetch_assoc($res);
-                        echo   $metodo["descripcion"] ;
+                        echo $metodo["descripcion"] ;
 
                     }
                     ?>
                 </p>
 
                 <?php
-                    if(isset($_GET["m"]))
+                    if (isset($_GET["m"]) && $metodo["regresa_html"])
                     {
-
-                        if($metodo["regresa_html"]) echo "<h2>Regresa HTML </h2>";
-
+                        echo "<h2>Regresa HTML </h2>";
                     }
-                ?>
 
-                <?php
                     if(isset($_GET["m"])){
                         $res = mysql_query("select * from metodo where id_metodo = " . $_GET["m"]) or die(mysql_error());
                         $metodo = mysql_fetch_assoc($res);
 
-
-
                         $argsq = mysql_query("select * from argumento where id_metodo = ". $metodo["id_metodo"] ." order by ahuevo desc,nombre asc;") or die(mysql_error());
-
 
                         ?>
                         <h2>Argumentos</h2>
                         <table class="methods" style="margin-left:0; width:100%">
                         <tr>
-                            <th style="border:1px solid #ddd;">
-                                Nombre
-                            </th>
-                            <th style="border:1px solid #ddd;">
-                                Tipo
-                            </th>
-                            <th style="border:1px solid #ddd;">
-                                Default
-                            </th>
-                            <th style="border:1px solid #ddd;">
-                                Descripcion
-                            </th>
+                            <th style="border:1px solid #ddd;"> Nombre </th>
+                            <th style="border:1px solid #ddd;"> Tipo </th>
+                            <th style="border:1px solid #ddd;"> Default </th>
+                            <th style="border:1px solid #ddd;"> Descripcion </th>
                         </tr>
 
                         <?php
@@ -569,20 +220,16 @@
 
         </div>
     </div>
+
     <div class="footer">
         <div class="content">
             <div class="copyright">
-                x © 2011
+                 © 2011
             </div>
             <div class="links">
                             <a href="build_bd.php?project=<?php echo ( isset($_GET["project"]) && is_numeric($_GET["project"]) )? $_GET["project"] : "null" ; ?>">Respaldar Base de Datos</a>
             </div>
         </div>
-    </div>
-    <div id="fb-root">
-    </div>
-    <input type="hidden" autocomplete="off" id="post_form_id" name="post_form_id" value="d8f38124ed9e31ef3947198c6d26bff1"/>
-    <div id="fb-root">
     </div>
 </div>
 </body>
