+++ apigen/www/top.inc.php
@@ -1,0 +1,20 @@
+
+<?php
+    $_GET["project"] = 1;
+    require_once("../server/bootstrap.php");
+?>
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml" lang="en" >
+<head>
+<script src="//code.jquery.com/jquery-1.12.4.min.js" integrity="sha256-ZosEbRLbNQzLpnKIkEdrPv7lOy9C27hHQ+Xp8a4MxAQ=" crossorigin="anonymous"></script>
+<title>Web Framework</title>
+
+<script src="js.js"></script>
+
+<link type="text/css" rel="stylesheet" href="media/f.css"/>
+
+</head>
+<body class="safari4 mac Locale_en_US">
+<input type="hidden" autocomplete="off" id="post_form_id" name="post_form_id" value="d8f38124ed9e31ef3947198c6d26bff1"/>
+<div id="FB_HiddenContainer" style="position:absolute; top:-10000px; width:0px; height:0px;">
+</div>
